requirements.txt
flask
flask-jwt-extended
requests
PyJWT
python-dotenv